remove cruft
